Fix: Resolve module loading errors and finalize refactor.
This commit addresses a critical JavaScript module loading error that occurred after the major refactoring, which was causing a blank screen on deployment.

Key changes in this commit:
- Corrected function name mismatches between modules. The `setupProfessorDashboard` and `setupStudentDashboard` functions were renamed to `initProfessorDashboard` and `initStudentDashboard` respectively in their defining files (`professor.js`, `student.js`) and the calling module (`auth.js`) to resolve the `does not provide an export named` error.

This commit builds upon the previous full-stack refactoring, which included:
- Correcting the main script import in `public/index.html`.
- Fortifying backend Cloud Functions with improved error handling.
- Implementing robust, error-proof data fetching on the frontend.
- Fixing UI bugs in both professor and student dashboards.
- Modularizing the frontend into `auth.js`, `professor.js`, `student.js`, etc.
- Cleaning up the project structure and optimizing dependencies.
- Enhancing the CI/CD workflow with caching.
