+++ public/index.html
@@ -200,6 +200,6 @@
     </div>
 
     <!-- The main application script will now handle all logic -->
-    <script type="module" src="/js/main.js"></script>
+    <script type="module" src="/js/firebase-init.js"></script>
 </body>
 </html>
